requirements.txt
fastapi
google-cloud-speech
python-dotenv
uvicorn
aiofiles
gtts
playsound
